requirements.txt
python-telegram-bot[webhooks]~=20.0
python-dotenv~=0.21.0
pymongo~=4.3.3